Rename files to swe_flux_{ceed,petsc}.h
The "petsc" is somewhat of a misnomer meaning "just PETSc" CPU code for
the fluxes.
